# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NC3 Testing Platform API",
    "summary": "v4.0 backend MVP for the NC3 Testing Platform.",
    "version": "4.0.1"
  },
  "paths": {
...
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "List notifications",
        "description": "The caller's inbox, newest first.",
        "operationId": "list_notifications_api_v1_notifications_get",
        "security": [
          {
            "OpenIdConnect": []
          },
          {
            "ApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor returned as `next_cursor` by the previous page.",
              "title": "Cursor"
            },
            "description": "Opaque cursor returned as `next_cursor` by the previous page."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Max items per page.",
              "default": 50,
              "title": "Limit"
            },
            "description": "Max items per page."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Notification_"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/read-all": {
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Mark all as read",
        "description": "Mark every unread notification belonging to the caller.\n\nReturns no body: the caller already knows the outcome, and re-sending the\ninbox here would duplicate `GET /notifications`.",
        "operationId": "mark_all_read_api_v1_notifications_read_all_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "OpenIdConnect": []
          },
          {
            "ApiKey": []
          }
        ]
      }
    },
    "/api/v1/notifications/{notification_id}/read": {
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Mark one as read",
        "description": "Set `read_at` on one of the caller's notifications.",
        "operationId": "mark_read_api_v1_notifications__notification_id__read_post",
        "security": [
          {
            "OpenIdConnect": []
          },
          {
            "ApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid7",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/webhook": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "Get the organization webhook",
        "description": "The organization's SIEM endpoint. `404` when none is configured.",
        "operationId": "get_webhook_api_v1_notifications_webhook_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationWebhook"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "OpenIdConnect": []
          },
          {
            "ApiKey": []
          }
        ]
      },
      "put": {
        "tags": [
          "notifications"
        ],
        "summary": "Create or replace the organization webhook",
        "description": "Set the single webhook configuration for the organization.",
        "operationId": "upsert_webhook_api_v1_notifications_webhook_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationWebhookUpsert"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationWebhook"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "OpenIdConnect": []
          },
          {
            "ApiKey": []
          }
        ]
      },
      "delete": {
        "tags": [
          "notifications"
        ],
        "summary": "Disable the organization webhook",
        "description": "Delete the configuration, which disables the integration.\n\nA `DELETE` rather than a revoke: nothing about the configuration needs to\nstay visible after it is switched off.",
        "operationId": "delete_webhook_api_v1_notifications_webhook_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "OpenIdConnect": []
          },
          {
            "ApiKey": []
          }
        ]
      }
    },
    "/api/v1/notifications/{notification_id}": {
      "delete": {
        "tags": [
          "notifications"
        ],
        "summary": "Dismiss a notification",
        "description": "Permanently remove the caller's row.\n\nDismissal is deletion in v4.0 — there is no archived state, and no other user\nis affected because the row was never shared.",
        "operationId": "dismiss_notification_api_v1_notifications__notification_id__delete",
        "security": [
          {
            "OpenIdConnect": []
          },
          {
            "ApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid7",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/account": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Get the current account",
        "description": "The caller's `app_user` projection: identity, organization, role, preference.",
        "operationId": "get_account_api_v1_account_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "OpenIdConnect": []
          },
          {
            "ApiKey": []
          }
        ]
      },
      "patch": {
        "tags": [
          "account"
        ],
        "summary": "Update account preferences",
        "description": "Change the email-notification opt-in.\n\nProfile fields are not editable here. They live in the identity provider and reach this\nprojection through claim updates.",
        "operationId": "update_account_api_v1_account_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "OpenIdConnect": []
          },
          {
            "ApiKey": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Account": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid7",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid7",
            "title": "Organization Id"
          },
          "organization_role": {
            "$ref": "#/components/schemas/OrganizationRole",
            "description": "Role within the organization. Platform-administrator status is a separate identity-provider claim and never appears here."
          },
          "email_notifications_enabled": {
            "type": "boolean",
            "title": "Email Notifications Enabled"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "organization_id",
          "organization_role",
          "email_notifications_enabled"
        ],
        "title": "Account",
        "description": "The caller's `app_user` projection.\n\nRead-only. Identity, credentials, display name, and email are owned by the identity provider\nand edited there; they arrive here through claim updates. The one field this\nplatform owns is the email-notification preference."
      },
      "AccountUpdate": {
        "properties": {
          "email_notifications_enabled": {
            "type": "boolean",
            "title": "Email Notifications Enabled"
          }
        },
        "type": "object",
        "required": [
          "email_notifications_enabled"
        ],
        "title": "AccountUpdate",
        "description": "The only account field this API can change."
      },
...
      "FieldError": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Dotted path to the offending field, e.g. `body.email`."
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "name",
          "reason"
        ],
        "title": "FieldError",
        "description": "One field-level validation error (RFC 9457 extension member)."
      },
...
      "Notification": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid7",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Namespaced notification type. v4.0 covers verification completion, regressions, scan completion and failure, retention warnings, and token expiry. The vocabulary is code-owned and extends freely.",
            "examples": [
              "scan.completed"
            ]
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "description": "Version of this type's `data` shape.",
            "examples": [
              "1.0"
            ]
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data",
            "description": "Type-specific payload."
          },
          "read_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read At",
            "description": "Null while unread. There is no separate flag."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "schema_version",
          "created_at"
        ],
        "title": "Notification",
        "description": "One inbox item belonging to one user."
      },
      "OrganizationRole": {
        "type": "string",
        "enum": [
          "member",
          "organization_admin"
        ],
        "title": "OrganizationRole",
        "description": "Role within one organization."
      },
      "OrganizationWebhook": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid7",
            "title": "Id"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid7",
            "title": "Organization Id"
          },
          "endpoint_url": {
            "type": "string",
            "pattern": "^https://",
            "format": "uri",
            "title": "Endpoint Url"
          },
          "created_by_user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid7"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By User Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "organization_id",
          "endpoint_url",
          "created_at",
          "updated_at"
        ],
        "title": "OrganizationWebhook",
        "description": "The organization's single SIEM integration endpoint.\n\nThe signing secret is never returned. It is set once on write and thereafter\nonly ever used to sign outgoing payloads."
      },
      "OrganizationWebhookUpsert": {
        "properties": {
          "endpoint_url": {
            "type": "string",
            "pattern": "^https://",
            "format": "uri",
            "title": "Endpoint Url"
          },
          "signing_secret": {
            "type": "string",
            "minLength": 32,
            "title": "Signing Secret",
            "description": "Shared secret used to sign delivered payloads. Stored encrypted and never returned. The payload's own `schema_version` belongs to the signed contract, not to this configuration."
          }
        },
        "type": "object",
        "required": [
          "endpoint_url",
          "signing_secret"
        ],
        "title": "OrganizationWebhookUpsert",
        "description": "Create or replace the webhook configuration.\n\nA `PUT` rather than a `POST`, because an organization has zero or one of these"
      },
...
      "Page_Notification_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Notification"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "Cursor for the next page; `null` when there are no more results."
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "Page[Notification]"
      },
...
      "ProblemDetail": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "URI reference identifying the problem type.",
            "default": "about:blank"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Short, human-readable summary of the problem type."
          },
          "status": {
            "type": "integer",
            "title": "Status",
            "description": "HTTP status code."
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "Human-readable explanation for this occurrence."
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance",
            "description": "URI reference identifying this occurrence."
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors",
            "description": "Field-level validation errors (extension)."
          }
        },
        "type": "object",
        "required": [
          "title",
          "status"
        ],
        "title": "ProblemDetail",
        "description": "RFC 9457 problem detail."
      },
...
    },
    "securitySchemes": {
      "OpenIdConnect": {
        "type": "openIdConnect",
        "description": "OpenID Connect token issued by the platform identity provider. Some operations additionally require current MFA assurance, read from the token at request time.",
        "openIdConnectUrl": "https://idp.example.invalid/.well-known/openid-configuration"
      },
      "ApiKey": {
        "type": "apiKey",
        "description": "Platform API key: `Authorization: Bearer <key>`. Scopes are `read_only` and `full_scan`. A scan launched with a key is recorded with `source = api`. Creating or revoking a key requires an OpenID Connect token with current MFA assurance.",
        "in": "header",
        "name": "Authorization"
      }
    }
  }
}